requirements.txt
Flask>=2.0
google-generativeai>=0.5.0
